feat(world_noaa_ghcn): add recurring Prefect pipeline
NCEI rewrites the current year's archive daily and reconstructs the whole
archive weekly, so this runs weekly and rebuilds a trailing window of
year-partitions rather than all 3.19 billion rows.

Two properties of the source shape the design.

File modification times cannot detect change: the weekly reconstruction
re-stamps all 264 by_year files with one identical timestamp, so an mtime diff
marks everything changed on every run. The guard is the source's max coverage
date, read from the data itself, compared against the registered coverage.

Only recent years move, and the table is partitioned by year. Uploading with
dump_mode="append" replaces just the window's blobs and leaves the other ~260
untouched, then dbt rebuilds the table from all of them. dump_mode="overwrite"
would delete the staging table and every blob under it, destroying the entire
backfill — a comment says so at the call site.

A trailing window cannot follow the reconstruction of *old* years, so
full_refresh rebuilds all 264 partitions. It is a multi-hour run, deliberately
not scheduled.

The transform is not duplicated. constants.py and clean.py move from
models/world_noaa_ghcn/code/ into the pipeline package, and the onboarding
scripts re-export them, so the one-shot backfill and the recurring refresh run
byte-identical code rather than two copies free to drift.

Coverage is AllFree for every table. GHCN-Daily is a US federal work in the
public domain and the dataset shipped fully open; a BD Pro rolling window would
retroactively close data that is already public, so that stays a decision to
make explicitly rather than a side effect of adding a pipeline.

Verified locally: flow imports, deploy_flows discovers world_noaa_ghcn_flow (a
silent no-deploy being the trap here), the refresh window resolves correctly
across a year boundary, and clean_all end-to-end on 1763 reproduces the exact
onboarding row counts. The inventory came back 44 rows larger, which is real
source movement — the live file grew 2,024 bytes overnight — not a regression.

job_variables sets memory_limit as well as memory; the work pool's job template
has no `memory` key, so a flow that sets only that silently gets 4Gi.
